refactor: extract add-to-dictionary action and fix tray style filtering
- Extract addSelectedTextToDictionary into a shared action with user
  feedback (toast when no text is selected, snackbar on error).
- Remove 'Add Selection to Dictionary' from the tray context menu.
- Fix tray Styles submenu to only show currently enabled/active tones
  using getActiveManualToneIds instead of all available tones.
